task/subtask stuff

+++ backend/tasks/util.dataLoadHelpers.coffee
@@ -29,7 +29,9 @@ DELETE =
 
 
 buildUniqueSubtaskName = (subtask, overrideBatchId) ->
-  parts = [overrideBatchId||subtask.batch_id, subtask.task_name, subtask.data.dataType]
+  parts = [overrideBatchId||subtask.batch_id, subtask.task_name]
+  if subtask.data.dataType && !subtask.task_name.endsWith(subtask.data.dataType)
+    parts.push(subtask.data.dataType)
   if subtask.data.rawTableSuffix
     parts.push(subtask.data.rawTableSuffix)
   parts.join('_')
@@ -65,7 +67,7 @@ recordChangeCounts = (subtask, opts={}) -> Promise.try () ->
 
   subid = buildUniqueSubtaskName(subtask)
   subset =
-    data_source_id: subtask.task_name
+    data_source_id: opts.data_source_id || subtask.task_name
   _.extend(subset, subtask.data.subset)
 
   dbs.transaction 'normalized', (transaction) ->
@@ -74,15 +76,15 @@ recordChangeCounts = (subtask, opts={}) -> Promise.try () ->
       if subtask.data.deletes == DELETE.UNTOUCHED
         # check if any rows will be left active after delete, and error if not; for efficiency, just grab the id of the
         # first such row rather than return all or count them all
-        tables.normalized[subtask.data.dataType](subid: subtask.data.normalSubid, transaction: transaction)
+        q = tables.normalized[subtask.data.dataType](subid: subtask.data.normalSubid, transaction: transaction)
         .select('rm_raw_id')
         .where(batch_id: subtask.batch_id)
         .where(subset)
         .whereNull('deleted')
         .limit(1)
-        .then (row) ->
+        q.then (row) ->
           if !row?.length
-            throw new HardFail("operation would delete all active rows for #{subtask.task_name}")
+            throw new HardFail("operation would delete all active rows for #{subtask.task_name}: #{q}")
         .then () ->
           # mark any rows not updated by this task (and not already marked) as deleted -- we only do this when doing a full
           # refresh of all data, because this would be overzealous if we're just doing an incremental update; the update
@@ -139,18 +141,18 @@ recordChangeCounts = (subtask, opts={}) -> Promise.try () ->
         # successfully commits for data safety
         dbs.transaction (mainDbTransaction) ->
           Promise.each results, (r) ->
-            markForDelete r.rm_property_id, subtask.task_name, subtask.batch_id,
+            markForDelete r.rm_property_id, (opts.data_source_id || subtask.task_name), subtask.batch_id,
               deletesTable: opts.deletesTable
               transaction: mainDbTransaction
 
 
 # this function flips inactive rows to active, active rows to inactive, and deletes now-inactive and extraneous rows
-activateNewData = (subtask, {tableProp, transaction, deletes} = {}) -> Promise.try () ->
+activateNewData = (subtask, {tableProp, transaction, deletes, skipIndicatedDeletes, data_source_id} = {}) -> Promise.try () ->
   logger.spawn(subtask.task_name).debug subtask
 
   tableProp ?= 'combined'
   subset =
-    data_source_id: subtask.task_name
+    data_source_id: data_source_id || subtask.task_name
   _.extend(subset, subtask.data.subset)
 
   # wrapping this in a transaction improves performance, since we're editing some rows twice
@@ -179,7 +181,7 @@ activateNewData = (subtask, {tableProp, transaction, deletes} = {}) -> Promise.t
         .where
           data_source_id: dbs.get('main').raw("updater.data_source_id")
           rm_property_id: dbs.get('main').raw("updater.rm_property_id")
-          update_source: subtask.task_name
+          update_source: data_source_id || subtask.task_name
           batch_id: subtask.batch_id
           active: false
       .update(active: dbs.get('main').raw('NOT "active"'))
@@ -189,18 +191,20 @@ activateNewData = (subtask, {tableProp, transaction, deletes} = {}) -> Promise.t
       # delete inactive rows
       tables.finalized[tableProp](transaction: transaction)
       .where
-        data_source_id: subtask.task_name
+        data_source_id: data_source_id || subtask.task_name
         active: false
       .delete()
     .then () ->
+      if skipIndicatedDeletes
+        return
       # delete rows marked explicitly for deletion
       tables.finalized[tableProp](transaction: transaction, as: 'deleter')
-      .where(data_source_id: subtask.task_name)
+      .where(data_source_id: data_source_id || subtask.task_name)
       .whereExists () ->
         tables.deletes[tableProp](transaction: this)
         .select(1)
         .where
-          data_source_id: subtask.task_name
+          data_source_id: data_source_id || subtask.task_name
           batch_id: subtask.batch_id
           rm_property_id: dbs.get('main').raw("deleter.rm_property_id")
       .delete()
@@ -208,7 +212,7 @@ activateNewData = (subtask, {tableProp, transaction, deletes} = {}) -> Promise.t
         # clean up after itself in the deletes table
         tables.deletes[tableProp](transaction: transaction)
         .where
-          data_source_id: subtask.task_name
+          data_source_id: data_source_id || subtask.task_name
           batch_id: subtask.batch_id
         .delete()
     .then () ->
@@ -348,9 +352,10 @@ normalizeData = (subtask, options) -> Promise.try () ->
           dataType: subtask.data.dataType
           subid: subtask.data.normalSubid
           dataSourceType: options.dataSourceType
+          idField: options.idField || 'rm_property_id'
         })
-        .then (rm_property_id) ->
-          successes.push(rm_property_id)
+        .then (id) ->
+          successes.push(id)
         .catch analyzeValue.isKnexError, (err) ->
           jsonData = util.inspect(updateRow, depth: null)
           tables.temp(subid: rawSubid)
@@ -449,7 +454,7 @@ updateRecord = (opts) -> Promise.try () ->
         data_source_id: updateRow.data_source_id
       .update(updateRow)
   .then () ->
-    updateRow.rm_property_id
+    updateRow[opts.idField]
 
 
 getValues = (list, target) ->
